# backend/api/rest/types.gen.go
// Package rest provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.13.4 DO NOT EDIT.
package rest

// Expense defines model for Expense.
type Expense struct {
	Id     *int    `json:"id,omitempty"`
	Price  *int    `json:"price,omitempty"`
	Title  *string `json:"title,omitempty"`
	UserId *int    `json:"userId,omitempty"`
}

// Expenses defines model for Expenses.
type Expenses = []Expense

// NotFound defines model for NotFound.
type NotFound struct {
	Code   *string `json:"code,omitempty"`
	Errors *[]struct {
		Code  *string `json:"code,omitempty"`
		Field *string `json:"field,omitempty"`
	} `json:"errors,omitempty"`
	Message *string `json:"message,omitempty"`
	Type    *string `json:"type,omitempty"`
}

// ServiceUnavailable defines model for ServiceUnavailable.
type ServiceUnavailable struct {
	Code   *string `json:"code,omitempty"`
	Errors *[]struct {
		Code  *string `json:"code,omitempty"`
		Field *string `json:"field,omitempty"`
	} `json:"errors,omitempty"`
	Message *string `json:"message,omitempty"`
	Type    *string `json:"type,omitempty"`
}

// User defines model for User.
type User struct {
	Id   *int    `json:"id,omitempty"`
	Name *string `json:"name,omitempty"`
}

// Users defines model for Users.
type Users = []User

// N404NotFound defines model for 404_NotFound.
type N404NotFound = NotFound

// N503ServiceUnavailable defines model for 503_ServiceUnavailable.
type N503ServiceUnavailable = ServiceUnavailable

// FindUser defines model for FindUser.
type FindUser struct {
	UserId *int `json:"userId,omitempty"`
}

// NewExpense defines model for NewExpense.
type NewExpense struct {
	Price  *int    `json:"price,omitempty"`
	Title  *string `json:"title,omitempty"`
	UserId *int    `json:"userId,omitempty"`
}

// NewUser defines model for NewUser.
type NewUser struct {
	Name *string `json:"name,omitempty"`
}

// GetV1ExpensesJSONBody defines parameters for GetV1Expenses.
type GetV1ExpensesJSONBody struct {
	UserId *int `json:"userId,omitempty"`
}

// PostV1ExpensesJSONBody defines parameters for PostV1Expenses.
type PostV1ExpensesJSONBody struct {
	Price  *int    `json:"price,omitempty"`
	Title  *string `json:"title,omitempty"`
	UserId *int    `json:"userId,omitempty"`
}

// DeleteV1ExpensesExpenseIdJSONBody defines parameters for DeleteV1ExpensesExpenseId.
type DeleteV1ExpensesExpenseIdJSONBody struct {
	UserId *int `json:"userId,omitempty"`
}

// GetV1ExpensesExpenseIdJSONBody defines parameters for GetV1ExpensesExpenseId.
type GetV1ExpensesExpenseIdJSONBody struct {
	UserId *int `json:"userId,omitempty"`
}

// PutV1ExpensesExpenseIdJSONBody defines parameters for PutV1ExpensesExpenseId.
type PutV1ExpensesExpenseIdJSONBody struct {
	Price  *int    `json:"price,omitempty"`
	Title  *string `json:"title,omitempty"`
	UserId *int    `json:"userId,omitempty"`
}

// PostV1UsersJSONBody defines parameters for PostV1Users.
type PostV1UsersJSONBody struct {
	Name *string `json:"name,omitempty"`
}

// PutV1UsersUserIdJSONBody defines parameters for PutV1UsersUserId.
type PutV1UsersUserIdJSONBody struct {
	Name *string `json:"name,omitempty"`
}

// GetV1ExpensesJSONRequestBody defines body for GetV1Expenses for application/json ContentType.
type GetV1ExpensesJSONRequestBody GetV1ExpensesJSONBody

// PostV1ExpensesJSONRequestBody defines body for PostV1Expenses for application/json ContentType.
type PostV1ExpensesJSONRequestBody PostV1ExpensesJSONBody

// DeleteV1ExpensesExpenseIdJSONRequestBody defines body for DeleteV1ExpensesExpenseId for application/json ContentType.
type DeleteV1ExpensesExpenseIdJSONRequestBody DeleteV1ExpensesExpenseIdJSONBody

// GetV1ExpensesExpenseIdJSONRequestBody defines body for GetV1ExpensesExpenseId for application/json ContentType.
type GetV1ExpensesExpenseIdJSONRequestBody GetV1ExpensesExpenseIdJSONBody

// PutV1ExpensesExpenseIdJSONRequestBody defines body for PutV1ExpensesExpenseId for application/json ContentType.
type PutV1ExpensesExpenseIdJSONRequestBody PutV1ExpensesExpenseIdJSONBody

// PostV1UsersJSONRequestBody defines body for PostV1Users for application/json ContentType.
type PostV1UsersJSONRequestBody PostV1UsersJSONBody

// PutV1UsersUserIdJSONRequestBody defines body for PutV1UsersUserId for application/json ContentType.
type PutV1UsersUserIdJSONRequestBody PutV1UsersUserIdJSONBody